requirements.txt
acres==0.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apify_client==1.12.0
apify_shared==1.4.1
asgiref==3.9.1
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.39.13
botocore==1.39.13
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cssselect==1.3.0
cssselect2==0.8.0
distlib==0.3.9
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.58.5
geographiclib==2.0
geopy==2.4.1
geotext==0.4.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jmespath==1.0.1
kombu==5.5.4
looseversion==1.3.0
lxml==6.0.0
lxml_html_clean==0.4.2
more-itertools==10.7.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.2
oauthlib==3.3.1
openai==1.97.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pbr==6.1.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
readability-lxml==0.8.4.1
redis==6.2.0
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stevedore==5.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.2.3.post1
